Move definition of video into separate module in domain
